Update to arch + forge port + keybinds
