+++ .gitignore
@@ -39,3 +39,4 @@ hs_err_*.log
 replay_*.log
 *.hprof
 *.jfr
+.architectury-transformer/debug.log
