+++ common/src/main/java/gg/meza/ConfigData.java
