+++ common/src/main/resources/assets/soundsbegone/icon.png
