+++ common/src/main/resources/assets/soundsbegone/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "soundsbegone.config" : "Open Config Screen",
+  "soundsbegone.keybinds" : "SoundsBeGone"
+}
