+++ common/src/main/resources/soundsbegone.accesswidener
@@ -1,0 +1,1 @@
+accessWidener   v2  named
