+++ common/src/main/resources/soundsbegone.client.mixins.json
@@ -1,0 +1,12 @@
+{
+	"required": true,
+	"package": "gg.meza.client.mixin",
+	"compatibilityLevel": "JAVA_17",
+	"minVersion": "0.8.4",
+	"client": [
+		"SoundInterceptorMixin"
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
