+++ fabric/src/main/java/gg/meza/fabric/ModMenuIntegration.java
@@ -1,0 +1,15 @@
+package gg.meza.fabric;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import gg.meza.client.ConfigScreen;
+
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return ConfigScreen::getConfigeScreen;
+    }
+
+
+}
