+++ forge/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "SoundsBeGone",
+    "pack_format": 13
+  }
+}
