+++ gradle.properties
@@ -1,19 +1,23 @@
 # Done to increase the memory available to gradle.
-org.gradle.jvmargs=-Xmx1G
+org.gradle.jvmargs=-Xmx2048M
 org.gradle.parallel=true
 
-# Fabric Properties
-# check these on https://fabricmc.net/develop
 minecraft_version=1.20.1
-yarn_mappings=1.20.1+build.8
-loader_version=0.14.21
+enabled_platforms=fabric,forge
 
-# Mod Properties
-mod_version=VERSION
-maven_group=gg.meza
 archives_base_name=soundsbegone
+mod_version=1.0.0
+maven_group=gg.meza
 
-# Dependencies
-fabric_version=0.84.0+1.20.1
+architectury_version=9.0.6
+yarn_mappings=1.20.1+build.8
+
+fabric_loader_version=0.14.21
+fabric_api_version=0.87.0+1.20.1
 modmenu_version=7.0.1
 cloth_version=11.0.99
+
+forge_version=1.20.1-47.1.3
+
+
+
