--- src/client/resources/soundsbegone.client.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "gg.meza.mixin.client",
-	"compatibilityLevel": "JAVA_17",
-	"client": [
-      "SoundInterceptorMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
