--- src/main/resources/soundsbegone.mixins.json
@@ -1,9 +1,0 @@
-{
-	"required": true,
-	"package": "gg.meza.mixin",
-	"compatibilityLevel": "JAVA_17",
-	"mixins": [],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
